+++ examples/test/const/test_const_typed_array_bad2d.pas
@@ -1,0 +1,5 @@
+program TestConstTypedArrBad2D;
+const
+    Grid: array[1..2, 1..2] of integer = (1, 2, 3, 4);
+begin
+end.