+++ examples/test/const/test_const_typed_array_badcount_fewer.pas
@@ -1,0 +1,5 @@
+program TestConstTypedArrBadFewer;
+const
+    Fib: array[1..3] of integer = (1, 1);
+begin
+end.