+++ examples/test/const/test_const_typed_array_badcount_more.pas
@@ -1,0 +1,5 @@
+program TestConstTypedArrBadMore;
+const
+    Fib: array[1..3] of integer = (1, 1, 2, 3);
+begin
+end.