+++ examples/test/const/test_const_typed_array_int.pas
@@ -1,0 +1,9 @@
+program TestConstTypedArrayInt;
+const
+    Fib: array[1..5] of integer = (1, 1, 2, 3, 5);
+var
+    i: integer;
+begin
+    for i := 1 to 5 do
+        writeln(Fib[i]);
+end.